feat: :sparkles: Added a PDF of my CV, updated a button's logo to indicate download, adjusted its functionality to initiate the CV download, and added a hover effect to lighten the button text color.
A PDF version of my CV was added to the project. The existing button's logo was replaced with a download symbol to better represent its purpose. The button's functionality was updated so that clicking it initiates the CV download. Additionally, a hover effect was implemented, changing the button's text color to a lighter shade of black for better user feedback and improved design consistency.